backend-api/requirements.txt
# Backend API Dependencies
# Okta AI Agent Demo - C2

# FastAPI Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Pydantic for data validation
pydantic==2.10.4
pydantic-settings==2.7.0

# HTTP Client
httpx==0.28.1

# Authentication
PyJWT==2.10.1
cryptography==44.0.0

# Anthropic Claude API
anthropic==0.42.0

# Utilities
python-dotenv==1.0.1

# Development
pytest==8.3.4
pytest-asyncio==0.25.2
